refactor(blocks-construct): revamp, code refactor, restructuring and modularised of Chat feature
