+++ src/features/chat/components/chat-search/chat-search.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { X } from 'lucide-react';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-import { mockChatContacts } from '../../data/chat.data';
+import { mockChatContacts } from '../../services/chat.services';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { ScrollArea } from '@/components/ui/scroll-area';
