+++ src/features/chat/components/chat/chat.tsx
@@ -1,16 +1,15 @@
 import { useState, useEffect } from 'react';
-import { Menu, ArrowLeft } from 'lucide-react';
-import { Button } from '@/components/ui/button';
 import { ChatStateContent } from '../chat-state-content/chat-state-content';
 import { ChatSidebar } from '../chat-sidebar/chat-sidebar';
 import { ChatSearch } from '../chat-search/chat-search';
 import { ChatUsers } from '../chat-users/chat-users';
 import { ChatContact } from '../../types/chat.types';
-import { mockChatContacts } from '../../data/chat.data';
+import { mockChatContacts } from '../../services/chat.services';
 import { Sheet, SheetContent } from '@/components/ui/sheet';
 import { ChatProfile } from '../chat-profile/chat-profile';
 import { Dialog, DialogContent, DialogHeader } from '@/components/ui/dialog';
 import { DialogDescription, DialogTitle } from '@radix-ui/react-dialog';
+import { ChatHeader } from '../chat-header/chat-header';
 
 function useIsMobile(breakpoint = 768) {
   const [isMobile, setIsMobile] = useState(() => window.innerWidth < breakpoint);
@@ -303,30 +302,3 @@ export const Chat = () => {
     </div>
   );
 };
-
-interface ChatHeaderProps {
-  selectedContact: ChatContact | null;
-  onMenuClick: () => void;
-  onInfoClick: () => void;
-  showBackButton: boolean;
-}
-
-const ChatHeader = ({
-  selectedContact,
-  onMenuClick,
-  showBackButton,
-}: Omit<ChatHeaderProps, 'onInfoClick'>) => (
-  <div className="flex items-center w-full min-h-[65px] px-4 border-b border-border">
-    <Button variant="ghost" size="icon" className="rounded-full" onClick={onMenuClick}>
-      {showBackButton ? (
-        <ArrowLeft className="w-6 h-6 text-medium-emphasis" />
-      ) : (
-        <Menu className="w-6 h-6 text-medium-emphasis" />
-      )}
-    </Button>
-
-    {selectedContact && (
-      <h2 className="lg:hidden text-lg font-semibold ml-3">{selectedContact.name}</h2>
-    )}
-  </div>
-);
