+++ src/features/chat/index.ts
@@ -8,3 +8,4 @@ export { ChatInput } from './components/chat-input/chat-input';
 export { ChatProfile } from './components/chat-profile/chat-profile';
 export { ForwardMessage } from './components/modals/forward-message/forward-message';
 export { EditGroupName } from './components/modals/edit-group-name/edit-group-name';
+export { ChatHeader } from './components/chat-header/chat-header';
