+++ src/pages/chat/chat.tsx
@@ -1,9 +1,9 @@
 import { Chat } from '@/features/chat';
 
-export function ChatPage() {
+export const ChatPage = () => {
   return (
     <div className="h-full w-full">
       <Chat />
     </div>
   );
-}
+};
