+++ src/pages/chat/index.ts
@@ -1,0 +1,1 @@
+export { ChatPage } from './chat';
